+++ chatroom/global.cc
@@ -2,7 +2,8 @@
 
 #include <QStyle>
 
-std::function<void(QWidget*)> repolish = [](QWidget* w) -> void {
+std::function<void(QWidget*)> repolish = [](QWidget* w) -> void
+{
   w->style()->unpolish(w);
   w->style()->polish(w);
 };