+++ chatroom/registerdialog.hpp
@@ -11,8 +11,12 @@
 #ifndef REGISTERDIALOG_HPP
 #define REGISTERDIALOG_HPP
 
+#include <QMap>
 #include <QDialog>
 #include <QString>
+#include <QJsonObject>
+
+#include "global.hpp"
 
 namespace Ui
 {
@@ -29,12 +33,15 @@ public:
 
 private slots:
   void on_verify_code_btn_clicked();
+  void slot_reg_mod_finish(QString str, ErrorCode err, ReqID id);
 
 private:
   void show_tip(const QString& str, bool ok);
+  void init_handlers();
 
 private:
   Ui::RegisterDialog* ui;
+  QMap<ReqID, std::function<void(const QJsonObject&)>> _handlers;
 };
 
 #endif  // REGISTERDIALOG_HPP